macOS: Inject storage handler into omnibar VC + extract Customize Responses menu row to its own file
