+++ macOS/DuckDuckGo/MainWindow/MainViewController.swift
@@ -332,7 +332,8 @@ final class MainViewController: NSViewController {
 
         aiChatOmnibarContainerViewController = AIChatOmnibarContainerViewController(
             themeManager: themeManager,
-            omnibarController: aiChatOmnibarController
+            omnibarController: aiChatOmnibarController,
+            duckAiNativeStorageHandler: NSApp.delegateTyped.duckAiNativeStorageHandler
         )
         aiChatOmnibarTextContainerViewController = AIChatOmnibarTextContainerViewController(
             omnibarController: aiChatOmnibarController,
